+++ CHANGELOG.md
@@ -46,11 +46,14 @@ New Features
 - Made COSMOSCatalog write an index parameter for both parameteric and real
   galaxy types to indicate the index of the object in the full COSMOS catalog.
   (#654, #694)
+- Added ability to specify lambda and r0 separately for Kolmogorov to have
+  GalSim do the conversion from radians to the given scale unit. (#657)
 - Changed `galsim.fits.writeMulti` to allow any of the "image"s to be
   already-built hdus, which are included as is. (#691)
 - Added optional `wcs` argument to `Image.resize()`. (#691)
 - Added `BaseDeviate.discard(n)` and `BaseDeviate.raw()`. (#691)
 - Added `sersic_prec` option to COSMOSCatalog.makeGalaxy(). (#691)
+- Added `convergence_threshold` as a parameter of HSMParams. (#709)
 
 
 Updates to galsim executable